requirements.txt
#latest version of openai works (1.31.0)
#some of the original embeddings done in openai(0.27.4)
openai==1.31.0
httpx==0.27.0
flask
python-dotenv==0.21.0
requests
flask-cors
anthropic